Moved Holder outside of CustomProp
